# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight local mock of the OpenAI REST API, covering Chat Completions\n(including streaming, tool calls, and multimodal content), Embeddings, Images,\nAudio, Models, Files, and Batch. Response shapes track the official OpenAI\nOpenAPI spec vendored at openai.yml; see memory.md for known deviations.",
        "title": "Mock OpenAI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists the model ids this mock recognizes. Note: chat/embeddings/images/audio endpoints accept any model string, not just these.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Models"
                ],
                "summary": "List models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/handlers.modelInfo"
                                    }
                                },
                                "object": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            }
        },
        "/models/{model_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a single model by id, if it's one of the ids returned by List models.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Models"
                ],
                "summary": "Get model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "model_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.modelInfo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errs.Detail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "param": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "errs.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/errs.Detail"
                }
            }
        },
        "handlers.modelInfo": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "owned_by": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token auth. Any non-empty token is accepted unless API_KEY is configured (see README). Example: \"Bearer sk-mock\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}